refactor: :truck: split button into subcrate
