+++ src/widgets/button/state.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+pub enum State {
+    #[default]
+    Normal,
+    Active,
+}
